move namespace from cupy.cuda to cupy.prof
