+++ cupy/prof/__init__.py
@@ -1,0 +1,3 @@
+from cupy.prof.timerange import timerange  # NOQA
+from cupy.prof.timerange import timerangeC  # NOQA
+from cupy.prof.timerange import TimeRangeDecorator  # NOQA
