+++ docs/source/cupy-reference/prof.rst
@@ -1,0 +1,9 @@
+Profiling
+=========
+
+time range
+----------
+
+.. autofunction:: cupy.prof.TimeRangeDecorator
+.. autofunction:: cupy.prof.timerange
+.. autofunction:: cupy.prof.timerangeC
